moved throttling to root folder
